Issue #154 refactor: RoutineLocalDataSource
