+++ app/src/main/java/com/lateinit/rightweight/data/datasource/RoutineLocalDataSourceImpl.kt
@@ -1,24 +1,17 @@
 package com.lateinit.rightweight.data.datasource
 
 import com.lateinit.rightweight.data.database.dao.RoutineDao
-import com.lateinit.rightweight.data.database.dao.SharedRoutineDao
 import com.lateinit.rightweight.data.database.entity.Day
 import com.lateinit.rightweight.data.database.entity.Exercise
 import com.lateinit.rightweight.data.database.entity.ExerciseSet
 import com.lateinit.rightweight.data.database.entity.Routine
-import com.lateinit.rightweight.data.database.entity.SharedRoutineDay
-import com.lateinit.rightweight.data.database.entity.SharedRoutineExercise
-import com.lateinit.rightweight.data.database.entity.SharedRoutineExerciseSet
 import com.lateinit.rightweight.data.database.intermediate.DayWithExercises
 import com.lateinit.rightweight.data.database.intermediate.RoutineWithDays
-import com.lateinit.rightweight.data.database.intermediate.SharedRoutineWithDays
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.emptyFlow
 import javax.inject.Inject
 
 class RoutineLocalDataSourceImpl @Inject constructor(
-    private val routineDao: RoutineDao,
-    private val sharedRoutineDao: SharedRoutineDao
+    private val routineDao: RoutineDao
 ) : RoutineLocalDataSource {
 
     override suspend fun insertRoutine(
@@ -68,25 +61,5 @@ class RoutineLocalDataSourceImpl @Inject constructor(
 
     override suspend fun removeRoutineById(routineId: String) {
         return routineDao.removeRoutineById(routineId)
-    }
-
-    override suspend fun insertSharedRoutineDetail(
-        days: List<SharedRoutineDay>,
-        exercises: List<SharedRoutineExercise>,
-        sets: List<SharedRoutineExerciseSet>
-    ) {
-       sharedRoutineDao.insertSharedRoutineDetail(days, exercises, sets)
-    }
-
-    override fun getSharedRoutineWithDaysByRoutineId(routineId: String): Flow<SharedRoutineWithDays> {
-        return sharedRoutineDao.getSharedRoutineWithDaysByRoutineId(routineId)
-    }
-
-    override fun getSelectedRoutine(routineId: String?): Flow<Routine> {
-        return if (routineId == null) {
-            emptyFlow()
-        } else {
-            routineDao.getSelectedRoutine(routineId)
-        }
     }
 }
